requirements.txt
# Google APIs for Gmail integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0

# Machine Learning and Text Processing
sentence-transformers>=2.2.2
numpy>=1.24.0

# HTTP requests
requests>=2.31.0 